pipeline/requirements.txt
apache-airflow-providers-postgres
apache-airflow-providers-redis
beautifulsoup4